+++ Route20/nsp_session_control_port_database.h
@@ -75,8 +75,8 @@ typedef struct
 	rtimer_t         *inactivityTimer;
 
 	void (*closeCallback)(uint16 locAddr);
-	void (*connectCallback)(decnet_address_t* remNode, uint16 locAddr, uint16 remAddr, byte *data, int dataLength);
-	void (*dataCallback)(uint16 locAddr, byte *data, int dataLength);
+	void (*connectCallback)(decnet_address_t* remNode, uint16 locAddr, uint16 remAddr, byte *data, byte dataLength);
+	void (*dataCallback)(uint16 locAddr, byte *data, uint16 dataLength);
 } session_control_port_t;
 
 void NspInitialiseScpDatabase(void);
